requirements.txt
Django==5.2.7
python-dotenv==1.1.0
django_vite_plugin==4.1.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0